requirements.txt
backoff==1.10.0
camelot-py==0.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
distro==1.5.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Simple-GeoIP==0.2.4
gunicorn==20.1.0
idna==2.10
ipaddress==1.0.23
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
opencv-python==4.5.2.52
openpyxl==3.0.7
pandas==1.2.4
pdfminer.six==20201018
pycparser==2.20
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
simple-geoip==0.1.1
six==1.16.0
sortedcontainers==2.4.0
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3
